requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
Pillow>=10.3.0
tensorflow>=2.16.0
